refactor(exporter): split config_test.go (4008 LoC) by subject (PR-2/5)
Mechanical decomposition of the 4008-line monolith into 8 subject-themed
sibling files + a slimmed residual. Pure file-level split — zero test
logic changes; all 116 tests preserved (verified via grep -cE '^func
Test' before/after = 116/116).

Layout (each file <800 LoC budget honored):

  config_resolve_test.go      282  Resolve / ResolveAt / StateFilter / ParseMetricKey
  config_dimensional_test.go  485  exact + regex dimensional labels
  config_three_state_test.go  530  ScheduledValue + UTC time windows + cross-midnight
  config_silent_mode_test.go  764  _silent_mode + _state_maintenance + expiry handling
  config_metadata_test.go     127  _metadata + ValidateTenantKeys reserved-key
  config_routing_test.go      429  _routing_profiles + receiver type + matcher dedup
  config_loaddir_test.go      381  LoadDir directory-mode + boundary enforcement
  config_incremental_test.go  670  IncrementalLoad + scanDirFileHashes + cache invariants
  config_test.go              461  residual: SV/SVScheduled/writeTestFile helpers +
                                   ConfigManagerBasics + UtilityFunctions +
                                   ConfigSourceDetection + WatchLoop_Integration

Originally proposed 7 files in the v2.8.0 refactor plan; bumped to 9
because residual would have been ~1500 LoC if loader-domain stayed
together — split DirectoryLoading and IncrementalReloading apart to
keep every file under the 800-line review budget.

Imports trimmed per file (each file imports only what it uses) — no
unused-import lint hits. Helpers (SV / SVScheduled / writeTestFile) stay
in config_test.go and are accessible to all sibling _test.go files via
Go's same-package test convention.

Verification:
  - go build -buildvcs=false ./... clean
  - go vet -buildvcs=false ./... clean
  - go test -buildvcs=false -race -count=1 ./... — all 9 packages pass
  - Spot-checks via -run flag confirm tests now execute from the new
    files (e.g. TestResolve_DimensionalBasic from config_dimensional_test.go)

Sets up the seam for PR-3 (diffAndReload split) and PR-4 (collapse
pkg/config duplication) — those will need to touch config_*_test.go;
having subject-themed files makes the touched-file footprint
predictable per PR.
